requirements.txt
# Core dependencies
streamlit>=1.29.0
pandas==2.1.3
numpy>=1.26.2
requests==2.31.0
boto3==1.34.0

# Visualization
plotly>=5.18.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
python-dotenv>=1.0.0
scipy==1.11.4

# Optional development tools
# black>=23.0.0
# pylint>=2.17.0
# pytest>=7.4.0
